# Optimize `always_redraw()` by reducing `Mobject` copying in `Mobject.become()`